requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
fsspec==2026.1
ftfy==6.3.1
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
open_clip_torch==3.2.0
packaging==26.0
pgvector==0.4.2
pillow==12.1.0
psycopg==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.2
rignore==0.7.6
safetensors==0.7.0
sentry-sdk==2.51.0
setuptools==80.10.2
shellingham==1.5.4
SQLAlchemy==2.0.46
sqlmodel==0.0.31
starlette==0.50.0
sympy==1.14.0
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
# uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.5.3
websockets==16.0

sentence_transformers
chromadb
pandas